+++ CHANGES.md
@@ -66,9 +66,10 @@ Changes in 13.0
 - HTML add "Calendars" fieldset in Configuration.php
 - Fix GitHub issue 305 alignment when a course has only a letter grade OR only a percent grade in Transcripts.fnc.php, thanks to @threehappypenguins
 - JS update Chart.js to version 4.5.1 in assets/js/Chart.js/
-- JS add Previous Next Student plugin to core in assets/js/jquery/previous-next-student/, warehouse.js & Gruntfile.js
+- JS add Previous Next Student plugin to core in assets/js/jquery-previous-next-student/, warehouse.js & Gruntfile.js
 - Remove List of words ignored during search operations in ListOutput.fnc.php
 - JS deprecate LO_search() function in warehouse.js, csp/functions.js & ListOutput.fnc.php
+- JS add Instant List Search Sorting plugin to core in assets/js/jquery-instant-list-search-sorting/, warehouse.js & Gruntfile.js
 
 Changes in 12.9.3
 -----------------